Fix rust tests
